+++ resources/ReleaseNotes.md
@@ -2,3 +2,5 @@
 
 - Simplify how columns are filtered in loading science and I-ALiRT
 - Make `mag.process.Calibration` more flexible to custom variable names
+- Add property `Harness` in `mag.meta.Science` to describe sensor harness
+- Fix issue with loading Excel files containing sensor meta data